Added default 'db' configuration
Defines a default adapter using an in-memory SQLite database. This is
done to prevent failurs when adding zend-db to an application that retrieves
the adapter via the container:

```php
$container->get(\Zend\Db\Adapter\AdapterInterface::class)
```

This was observed when adding ZendDeveloperTools to a ZF3 application
while also installing zend-db, but without first configuring zend-db.

Reported at: https://github.com/zendframework/tutorials/pull/57