Updated disconnect for persistent connections
The db2_pconnect allows persistent connections to be made in the connect function, but the the db2_close function in the disconnect function does not close those connections.  This change allows a boolean to be passed to the disconnect function with a value of true that will allow a persistent connection to be closed.